# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RestaurantType": {
				"enum": [
					"CAFE",
					"RESTAURANT",
					"BOTH"
				],
				"type": "string"
			},
			"RestaurantValidation": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/RestaurantType"
					},
					"address": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"type",
					"address",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRestaurantValidation": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/RestaurantType"
					},
					"address": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"type",
					"address",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RestaurantEnitity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/RestaurantType"
					},
					"address": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/CategoryEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"email",
					"type",
					"address",
					"password",
					"categories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"restaurant": {
						"$ref": "#/components/schemas/RestaurantEnitity"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/ProductEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"description",
					"restaurant",
					"products"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"stock": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/CategoryEntity"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"description",
					"price",
					"stock",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductValidation": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"stock": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"price",
					"stock",
					"category_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategory": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"restaurant_id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"restaurant_id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "simpleapiusingtsoa",
		"version": "1.0.0",
		"contact": {
			"name": "NirmalNyaupane"
		}
	},
	"paths": {
		"/restaurant": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string"
										},
										"createdAt": {
											"type": "string"
										},
										"id": {
											"type": "string"
										},
										"address": {
											"type": "string"
										},
										"type": {
											"$ref": "#/components/schemas/RestaurantType"
										},
										"email": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"id",
										"address",
										"type",
										"email",
										"name"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RestaurantValidation"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RestaurantEnitity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": []
			}
		},
		"/restaurant/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRestaurantValidation"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/product": {
			"post": {
				"operationId": "AddProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductValidation"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": []
			}
		},
		"/category": {
			"post": {
				"operationId": "Add",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RestaurantEnitity"
								}
							}
						}
					}
				},
				"tags": [
					"category"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategory"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"category"
				],
				"security": [],
				"parameters": []
			}
		},
		"/category/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}